saas_management_backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.3
pydantic-settings==2.2.1
# Database
sqlalchemy==2.0.27
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9 # For synchronous migration scripts
# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
email-validator==2.1.1
qdrant-client==1.7.0
pypdf==4.0.1
langchain-text-splitters==0.0.1
openai==1.13.3
python-multipart==0.0.9 # For file uploads